fix(downloads): resolve image download vanish bugs and lint errors
- Call hydrateDownloadStore on app mount so in-progress downloads
  survive app kills and reappear correctly on restart
- Keep completed image rows during hydration (previously filtered out),
  surfacing them as 'processing' so JS finalization can re-run
- Add resumeImageDownload to re-finalize processing entries on restart,
  handling all three recovery states: already unzipped, zip in app dir,
  zip still in WorkManager temp location
- Add disk-existence guard in proceedWithDownload to skip re-download
  when model files already exist on disk
- Fix wireZipListeners error handler to set failed status in store so
  the entry stays visible for user action
- Fix canCancel in TextModelsTab to use isActiveStatus instead of only
  checking pending/running, covering retrying/waiting_for_network states
- Remove debug panel from DownloadManagerScreen
- Extract resumeImageDownload into imageDownloadResume.ts to stay under
  max-lines limit; refactor resolveMmProj helper to reduce complexity
- Add timing logs around moveCompletedDownload and unzip for diagnosis
