+++ src/screens/DownloadManagerScreen/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { View, Text, FlatList, TouchableOpacity, ScrollView } from 'react-native';
+import { View, Text, FlatList, TouchableOpacity } from 'react-native';
 import { SafeAreaView } from 'react-native-safe-area-context';
 import { useNavigation } from '@react-navigation/native';
 import Icon from 'react-native-vector-icons/Feather';
@@ -9,9 +9,6 @@ import { useTheme, useThemedStyles } from '../../theme';
 import { createStyles } from './styles';
 import { ActiveDownloadCard, CompletedDownloadCard, formatBytes } from './items';
 import { useDownloadManager } from './useDownloadManager';
-import { useDownloadStore } from '../../stores/downloadStore';
-import { hydrateDownloadStore } from '../../services/downloadHydration';
-import logger from '../../utils/logger';
 
 export const DownloadManagerScreen: React.FC = () => {
   const navigation = useNavigation();
@@ -91,83 +88,6 @@ export const DownloadManagerScreen: React.FC = () => {
                 </Card>
               )}
             </View>
-
-            {/* Debug Panel - DEV only */}
-            {__DEV__ && (
-              <View style={{ margin: 12, padding: 12, backgroundColor: '#1a1a2e', borderRadius: 8, borderWidth: 1, borderColor: '#ff6b6b' }}>
-                <Text style={{ color: '#ff6b6b', fontWeight: '600', marginBottom: 8, fontSize: 12 }}>DEBUG PANEL</Text>
-
-                <Text style={{ color: '#aaa', fontSize: 11, marginBottom: 6 }}>Simulate on first active download:</Text>
-
-                {activeItems.length === 0 && (
-                  <Text style={{ color: '#666', fontSize: 11, marginBottom: 6 }}>No active downloads - start one first</Text>
-                )}
-
-                {activeItems.length > 0 && (() => {
-                  const item = activeItems[0];
-                  const downloadId = item.downloadId!;
-                  const modelKey = item.modelKey!;
-                  return (
-                    <View style={{ gap: 6 }}>
-                      <Text style={{ color: '#888', fontSize: 10 }}>Target: {item.fileName.substring(0, 30)}...</Text>
-
-                      <TouchableOpacity
-                        style={{ backgroundColor: '#c0392b', padding: 8, borderRadius: 6 }}
-                        onPress={() => {
-                          logger.log('[DEBUG] Simulating finalization error for', downloadId);
-                          useDownloadStore.getState().setStatus(downloadId, 'failed', { message: 'Simulated: unzip failed - file corrupted', code: 'file_corrupted' });
-                        }}
-                      >
-                        <Text style={{ color: '#fff', fontSize: 11 }}>Simulate unzip/finalization failure</Text>
-                      </TouchableOpacity>
-
-                      <TouchableOpacity
-                        style={{ backgroundColor: '#e67e22', padding: 8, borderRadius: 6 }}
-                        onPress={() => {
-                          logger.log('[DEBUG] Simulating network error for', downloadId);
-                          useDownloadStore.getState().setStatus(downloadId, 'failed', { message: 'Simulated: network connection lost', code: 'network_lost' });
-                        }}
-                      >
-                        <Text style={{ color: '#fff', fontSize: 11 }}>Simulate network error</Text>
-                      </TouchableOpacity>
-
-                      <TouchableOpacity
-                        style={{ backgroundColor: '#8e44ad', padding: 8, borderRadius: 6 }}
-                        onPress={() => {
-                          logger.log('[DEBUG] Simulating stuck processing for', downloadId);
-                          useDownloadStore.getState().setProcessing(downloadId);
-                        }}
-                      >
-                        <Text style={{ color: '#fff', fontSize: 11 }}>Simulate stuck in processing</Text>
-                      </TouchableOpacity>
-
-                      <TouchableOpacity
-                        style={{ backgroundColor: '#2980b9', padding: 8, borderRadius: 6 }}
-                        onPress={() => {
-                          logger.log('[DEBUG] Simulating app restart - clearing store then hydrating');
-                          useDownloadStore.getState().setAll([]);
-                          hydrateDownloadStore().then(() => {
-                            logger.log('[DEBUG] Hydration complete. Store:', JSON.stringify(Object.keys(useDownloadStore.getState().downloads)));
-                          });
-                        }}
-                      >
-                        <Text style={{ color: '#fff', fontSize: 11 }}>Simulate app restart (clear + hydrate)</Text>
-                      </TouchableOpacity>
-
-                      <TouchableOpacity
-                        style={{ backgroundColor: '#27ae60', padding: 8, borderRadius: 6 }}
-                        onPress={() => {
-                          logger.log('[DEBUG] Simulating store wipe only (no hydration) for', modelKey);
-                          useDownloadStore.getState().setAll([]);
-                        }}
-                      >
-                        <Text style={{ color: '#fff', fontSize: 11 }}>Simulate store wipe (no hydration)</Text>
-                      </TouchableOpacity>
-                    </View>
-                  );
-                })()}
-              </View>
-            )}
 
             {/* Storage Info */}
             {completedItems.length > 0 && (
